+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## `1.1.1-109-SNAPSHOT`
+## `1.2.1-113-SNAPSHOT`
 
 Release date `UNRELEASED`
+
+## `1.2.0-112`
+
+Release date `2023-03-21`
 
 - (feat api): add type hint to `exchange/get-input-stream`
 - (fix api BREAKING): change `:report-callback` CSP option to `:report-handler`